+++ migrationsTimescale/20210709000001-add-user-project-role-unique-constraint.js
@@ -1,0 +1,13 @@
+'use strict'
+module.exports = {
+  up: async (queryInterface, Sequelize) => {
+    return await queryInterface.sequelize.query(
+      'ALTER TABLE user_project_roles ADD CONSTRAINT user_project_roles_unique UNIQUE (project_id, user_id, role_id);'
+    )
+  },
+  down: async (queryInterface, Sequelize) => {
+    return await queryInterface.sequelize.quert(
+      'ALTER TABLE user_project_roles DROP CONSTRAINT user_project_roles_unique;'
+    )
+  }
+}