+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## 1.0.23 (2021-06-??)
+## 1.0.23 (2021-07-??)
 
 ### Features
 * **internal:** Support for message queue on segment creation ([CE-749](https://jira.rfcx.org/browse/CE-749))
 * **core:** Add unique constraint to user_project_roles table
+* **core:** Add permission field for project listing
 
 ### Bug Fixes
 * **core:** Guest user can not create stream in a project ([CE-961](https://jira.rfcx.org/browse/CE-961))